🛠️ Fix lint issues
