+++ src/app/api/properties/suggestions/route.ts
@@ -14,6 +14,7 @@ export const GET = async (request: NextRequest) => {
     const suggestions = await propertyService.getSuggestions(query, 5);
 
     return NextResponse.json(suggestions);
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (error) {
     return NextResponse.json(
       { error: "Failed to fetch suggestions" },
